refactor(jsi): start the shared engine layer with the neutral GSD core
First step of the Apple-side extraction (branch 2a). Introduces
NativeScript/jsi as the platform-neutral home for the engine binding layer that
Android will also compile, and moves in the pieces that are already neutral.

Moved verbatim:
  ffi/objc/shared/NativeApiBackendConfig.h -> jsi/shared/
  ffi/objc/shared/Tasks.{h,cpp}            -> jsi/shared/

SignatureDispatchCore.h could not simply move: it is only half neutral. The
mechanism (FNV-1a hashing, SignatureCallKind, the prepared-invoker and
dispatch-entry types, the sorted-table lookup, the NS_DISABLE_GSD switch) is
common, but the rest walks the Objective-C metadata through
metagen::MDMetadataReader and metagen::MDTypeKind. Android's metadata is a
different binary format, so that half stays under ffi/objc. The neutral half is
now jsi/shared/SignatureHashing.h, which SignatureDispatchCore.h includes.

Every consumer includes these through their full "ffi/objc/shared/..." path, so
forwarding headers are left behind and no Apple include or source file changes.
The only other Apple change is the CMake path for Tasks.cpp.

Adds scripts/check_jsi_layer_neutral.sh, which fails if Objective-C or a .mm
file appears under NativeScript/jsi. This is not paranoia: `#import
<Foundation/Foundation.h>` compiles fine under -x c++ on macOS, so a green Apple
build does not prove the layer is Android-compilable, and the regression would
only surface much later on the Android side.

Verified: iOS builds and reports 713 specs / 0 failures, unchanged from the
parent branch; the shared header also compiles standalone with
`clang++ -x c++ -std=c++20`. No Android file is touched by this branch.

(cherry picked from commit 9d62ca31c1ea543eb647325c9b8b6655dc44a749)
