+++ NativeScript/jsi/shared/Tasks.h
@@ -1,0 +1,20 @@
+#ifndef Tasks_h
+#define Tasks_h
+
+#include <vector>
+#include <functional>
+
+namespace nativescript {
+
+class Tasks {
+public:
+    static void Register(std::function<void()> task);
+    static void Drain();
+    static void ClearTasks();
+private:
+    static std::vector<std::function<void()>> tasks_;
+};
+
+}
+
+#endif /* Tasks_h */
